+++ src/components/toast/Toast.vue
@@ -2,8 +2,10 @@
     <transition
         :enter-active-class="transition.enter"
         :leave-active-class="transition.leave">
-
-        <div class="toast" v-show="isActive" :class="type">
+        <div
+            v-show="isActive"
+            class="toast"
+            :class="[type, position]">
             <div v-html="message"></div>
         </div>
     </transition>
@@ -18,17 +20,6 @@
         data() {
             return {
                 newDuration: this.duration || config.defaultToastDuration
-            }
-        },
-        methods: {
-            /**
-             * Add component to the DOM with it's classes,
-             * called from the Mixin.
-             */
-            insertEl() {
-                this.parent.className = ''
-                this.parent.classList.add('notices', 'is-toast', this.position)
-                this.parent.appendChild(this.$el)
             }
         }
     }